notebooks/SK_3_2/requirement.txt
appnope==0.1.2
backcall==0.2.0
bcrypt==3.2.0
branca==0.4.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
cycler==0.11.0
debugpy @ file:///opt/concourse/worker/volumes/live/fb0af3c0-b0a4-42c5-46e7-6ce9c84da602/volume/debugpy_1637091827531/work
decorator==4.4.2
Elucidata-resources @ git+ssh://git@gitlab.sirris.be/elucidata/elucidata_resources.git@851ff3d42d5ea9f4d33936878b2e25c8ec691406
folium==0.12.1.post1
fonttools==4.30.0
idna==3.3
ipykernel @ file:///opt/concourse/worker/volumes/live/6f6caad4-5c02-4b4a-5243-1eece346c27b/volume/ipykernel_1633545433252/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///opt/concourse/worker/volumes/live/4c3cce71-ced8-42c4-43fc-7be66d2225a8/volume/ipython_1643818166775/work
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.3
joblib==1.1.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core @ file:///opt/concourse/worker/volumes/live/365b665b-3a8e-478f-5f79-ff4d9b0854bd/volume/jupyter_core_1636537252467/work
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.2
numpy==1.22.3
packaging==21.3
pandas==1.4.1
paramiko==2.9.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
progressbar==2.5
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycparser==2.21
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
PyNaCl==1.5.0
pyparsing==3.0.7
python-dateutil==2.8.1
python-dotenv==0.19.2
pytz==2021.3
pyzmq @ file:///opt/concourse/worker/volumes/live/dbbc086d-2905-496b-54ee-08109e9ced8d/volume/pyzmq_1638435006316/work
requests==2.27.1
scipy==1.8.0
scp==0.14.4
seaborn==0.11.2
six==1.15.0
tornado @ file:///opt/concourse/worker/volumes/live/2c1a63a2-006b-48ee-56b9-0cfe8b4927f9/volume/tornado_1606942321278/work
tqdm==4.63.0
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
urllib3==1.26.8
wcwidth==0.2.5
